Fix bug creating a file with the same name of an existing folder. When is created the folder with that name dissapears from the UI. Looks a client bug
